Dtype

+++ experiments/Trees/prepare_field_data.py
@@ -133,7 +133,7 @@ def create_records(crops, labels, box_index, savedir, chunk_size=1000):
             chunk_labels = None
         
         #resize crops
-        resized_crops = [resize(x, height, width).astype("int16") for x in chunk_crops]
+        resized_crops = [resize(x.astype("float32"), height, width) for x in chunk_crops]
         
         filename = "{}/{}_{}.tfrecord".format(savedir, basename, counter)
         write_tfrecord(filename=filename,
